requirements.txt
fastapi==0.63.0
python-telegram-bot==13.4
